pyproject.toml
[project]
name = "autonomous_sre_bot"
version = "0.1.0"
description = "autonomous-sre-bot using crewAI"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.105.0,<1.0.0",
    "python-dotenv>=1.0.0",
    "langfuse",
    "openlit==1.32.0",
    "transformers",
    "opentelemetry-sdk>=1.27.0,<2.0.0",
    "opentelemetry-api>=1.27.0,<2.0.0",
    "opentelemetry-exporter-otlp>=1.27.0,<2.0.0",
    # MCP integration dependencies
    "mcp>=1.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "databricks-sdk>=0.59.0",
    "crewai-tools[mcp]>=0.38.0",
    "pytest>=8.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
mcp = [
    "crewai-tools[mcp]>=0.15.0",
]

[project.scripts]
autonomous_sre_bot = "autonomous_sre_bot.main:run_incident_management"
run_crew = "autonomous_sre_bot.main:run_incident_management"
train = "autonomous_sre_bot.main:train_incident_crew"
replay = "autonomous_sre_bot.main:replay"
test = "autonomous_sre_bot.main:test_incident_management"
incident_management = "autonomous_sre_bot.main:run_incident_management"
# Self-healing crew scripts
self_heal = "autonomous_sre_bot.main_self_heal:main"
self_heal_daemon = "autonomous_sre_bot.daemon_self_heal:main"
test_self_heal = "autonomous_sre_bot.test_self_heal:main"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"
